Graph layout

+++ source-code/drawing/Point.h
@@ -3,9 +3,9 @@
 
 
 struct Point {
-    int x;
-    int y;
-    Point(int x, int y) : x(x), y(y) {}
+    double x;
+    double y;
+    Point(double x, double y) : x(x), y(y) {}
     Point(Point* p) : x(p->x), y(p->y) {}
 };
